requirements.txt
boto3==1.35.30
cartopy==0.23.0
dotenv
geopandas==0.13.2
imageio==2.35.1
ipykernel==6.29.5
ipywidgets
jupyter
line_profiler
matplotlib==3.9.2
numpy==1.26.4
papermill
Pillow==10.4.0
pyarrow==17.0.0
pytest==8.3.3
pyyaml==6.0.2
rasterio==1.3.11
requests==2.32.3
SciencePlots==2.1.1
scipy==1.14.1
scikit-learn==1.5.1
shapely==1.7.1
torch==2.4.1
torchvision==0.19.1
torchaudio>=2.0.0
torchsummary
tqdm==4.66.5
transformers==4.38.2
timm
einops==0.8.0
huggingface_hub==0.26.2
elevation==1.1.3
opencv-python==4.10.0.84
pandas==2.2.3
omegaconf
accelerate
wandb
safetensors==0.5.0
tensorflow
tiktoken==0.5.2
python-dotenv==0.19.2
watermark==2.5.0
umap-learn